// frontend/src/models/trading-models.ts
export enum TimeFrameEnum {
  DAY = "DAY",
  MINUTE = "MINUTE",
  HOUR = "HOUR",
  WEEK = "WEEK",
  MONTH = "MONTH",
}

export enum StrategyEnum {
  OBV = "OBV",
  // RSI = "RSI",
  // MACD = "MACD",
  // SMA = "SMA",
  // BOLLINGERBANDS = "BBANDS",
  // STOCHASTICOSCILLATOR = "STOCH",
  // ADX = "ADX",
  // CDLMORNINGSTAR = "MORNINGSTAR",
  // CDL3WHITESOLDIERS = "3WHITESOLDIERS",
  // CDL3BLACKCROWS = "3BLACKCROWS",
  // CDLIDENTICAL3CROWS = "IDENTICAL3CROWS",
  // CDLEVENINGSTAR = "EVENINGSTAR",
  // CDLCONCEALBABYSWALL = "CONCEALBABYSWALL",
  // CDL3LINESTRIKE = "3LINESTRIKE",
}

export enum DirectionEnum {
  BUY = "BUY",
  SELL = "SELL",
  HOLD = "HOLD",
  CLOSE_BUYS_AND_SELL = "CLOSE BUYS AND SELL",
  CLOSE_SELLS_AND_BUY = "CLOSE SELLS AND BUY",
}